Added: moved utils, schemas for request and response
