+++ konfuzio_sdk/bento/utils.py
@@ -1,0 +1,24 @@
+"""Utility functions usable across different containerized AIs."""
+from typing import Any
+
+from pydantic import PlainSerializer, PlainValidator, WithJsonSchema, errors
+from typing_extensions import Annotated
+
+
+def hex_bytes_validator(o: Any) -> bytes:
+    """
+    Custom validator to be able to correctly serialize and unserialize bytes.
+    See https://github.com/pydantic/pydantic/issues/3756#issuecomment-1654425270
+    """
+    if isinstance(o, bytes):
+        return o
+    elif isinstance(o, bytearray):
+        return bytes(o)
+    elif isinstance(o, str):
+        return bytes.fromhex(o)
+    raise errors.BytesError
+
+
+HexBytes = Annotated[
+    bytes, PlainValidator(hex_bytes_validator), PlainSerializer(lambda b: b.hex()), WithJsonSchema({'type': 'string'})
+]
